fix bug

+++ NodejsApp/controllers/device.controller.js
@@ -1,6 +1,10 @@
-const {dbRun, dbAll, db} = require('../models/database')
+const { dbRun, dbAll } = require('../models/database')
 const handler = require('./handler')
 const uniqueId = require('../utils/uniqueId')
+
+const util = require('util')
+const fs = require('fs')
+const unlink = util.promisify(fs.unlink.bind(fs))
 
 class Device {
     handleErrCreate = async (id) => {
@@ -33,7 +37,7 @@ class Device {
         )
         const tagParams = values.map((value) => Object.values(value)).flat()
 
-        return {proTagQuery, proTagParams, tagQuery, tagParams}
+        return { proTagQuery, proTagParams, tagQuery, tagParams }
     }
 
     getAll = function (req, res) {
@@ -58,13 +62,14 @@ class Device {
 
     create = (req, res) => {
         const id = uniqueId()
+        const { data, repNum: replicateNumber } = req.body
         const deviceQuery = 'INSERT INTO DEVICE (ID, name, description, protocolType) VALUES (?, ?, ?, ?)'
-        const deviceParams = [id, req.body.name, req.body.description, req.body.protocol.toUpperCase()]
+        const deviceParams = [id, data.name, data.description, data.protocol.toUpperCase()]
 
-        const protocolName = req.body.protocol.toUpperCase()
-        var protocolParams = Object.values(req.body.config)
+        const protocolName = data.protocol.toUpperCase()
+        var protocolParams = Object.values(data.config)
 
-        const tagList = req.body.tagList
+        const tagList = data.tagList
 
         handler(res, async () => {
             if (protocolParams.length === 0) {
@@ -75,8 +80,9 @@ class Device {
             protocolParams.push(id)
             const protocolQuery = `INSERT INTO ${protocolName} VALUES (${',?'.repeat(protocolParams.length).slice(1)})`
 
-            if (tagList.length !== 0) {
-                const {proTagQuery, proTagParams, tagQuery, tagParams} = this.setupTagSql(tagList, protocolName, id)
+            console.log(tagList)
+            if (tagList.length !== 0 && Object.keys(tagList[0]).length !== 0) {
+                const { proTagQuery, proTagParams, tagQuery, tagParams } = this.setupTagSql(tagList, protocolName, id)
 
                 try {
                     await Promise.all([
@@ -153,11 +159,11 @@ class Device {
         const tagList = req.body.tagList
 
         handler(res, async () => {
-            if (tagList.length !== 0) {
+            if (tagList.length !== 0 && Object.keys(tagList[0]).length !== 0) {
                 const deleteQuery = `DELETE FROM TAG WHERE deviceID = ?`
                 await dbRun(deleteQuery, id)
 
-                const {proTagQuery, proTagParams, tagQuery, tagParams} = this.setupTagSql(tagList, protocolName, id)
+                const { proTagQuery, proTagParams, tagQuery, tagParams } = this.setupTagSql(tagList, protocolName, id)
                 await Promise.all([
                     dbRun(deviceQuery, deviceParams),
                     dbRun(protocolQuery, protocolParams),
@@ -174,14 +180,18 @@ class Device {
     }
 
     drop = function (req, res) {
-        const deviceQuery = 'DELETE FROM DEVICE WHERE ID = ?'
-        const deviceParams = [req.params.id]
+        const deleteDeviceQuery = 'DELETE FROM DEVICE WHERE ID = ?'
+        const deviceID = req.params.id
 
         handler(res, async () => {
-            await dbRun(deviceQuery, deviceParams)
+            await dbRun(deleteDeviceQuery, [deviceID])
+
+            const files = fs.readdirSync('./customJSON').filter(fn => fn.slice(9, 17) === deviceID);
+            const unlinkPromises = files.map(file => unlink(file))
+            await Promise.all(unlinkPromises)
 
             res.json({
-                key: req.params.id,
+                key: deviceID,
             })
         })
     }
